pyproject.toml
[project]
name = "classification-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "jupyter>=1.0.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "joblib>=1.1.0",
    "liac-arff>=2.5.0",
    "xgboost>=1.6.0",
    "scikit-plot>=0.3.7",
    "plotly>=5.0.0",
    "optuna>=3.0.0",
    "flaml>=2.3.6",
    "autogluon-tabular>=1.4.0",
]
